rename: 컴포넌트명 오타 수정; Couponconter -> CouponCounter
